refactor: move interfaces | remove comments
